refactor: Moved Report and ReportType to different package - Moved from tson.reporter to tson.reporter.report
