+++ src/main/java/com/euph28/tson/context/keyword/RequestVariable.java
@@ -4,8 +4,8 @@ import com.euph28.tson.context.TSONContext;
 import com.euph28.tson.context.VariableType;
 import com.euph28.tson.core.keyword.Keyword;
 import com.euph28.tson.core.keyword.KeywordType;
-import com.euph28.tson.reporter.Report;
-import com.euph28.tson.reporter.ReportType;
+import com.euph28.tson.reporter.report.Report;
+import com.euph28.tson.reporter.report.ReportType;
 import com.euph28.tson.reporter.TSONReporter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
