+++ src/main/java/com/euph28/tson/reporter/ReportRetriever.java
@@ -1,5 +1,8 @@
 package com.euph28.tson.reporter;
 
+import com.euph28.tson.reporter.report.Report;
+import com.euph28.tson.reporter.report.ReportType;
+
 import java.util.List;
 
 /**
@@ -10,7 +13,7 @@ public interface ReportRetriever<ReportDataType> {
      * Retrieve the report of the current {@link TSONReporter} based on its contained data. This method is executed
      * recursively until the root generates an output
      *
-     * @param derivedReportType Highest {@link ReportType#severity} {@link ReportType} within the sub-reports
+     * @param derivedReportType Highest {@link ReportType#getSeverity()} {@link ReportType} within the sub-reports
      * @param report            Current report within the {@link TSONReporter}
      * @param subReportList     List of sub-reports within the {@link TSONReporter}
      * @param layer             The current layer of this report (layer starts at 0 with root report)
@@ -22,7 +25,7 @@ public interface ReportRetriever<ReportDataType> {
     /**
      * Filter to determine if a report should be used or ignored
      *
-     * @param derivedReportType Highest {@link ReportType#severity} {@link ReportType} within the sub-reports
+     * @param derivedReportType Highest {@link ReportType#getSeverity()} {@link ReportType} within the sub-reports
      * @param report            Current report within the {@link TSONReporter}
      * @return Return {@code true} if the report should be used for generation. Return {@code false} if report should be skipped
      */
