+++ src/main/java/com/euph28/tson/restclientinterface/keyword/KeywordSend.java
@@ -2,8 +2,8 @@ package com.euph28.tson.restclientinterface.keyword;
 
 import com.euph28.tson.context.TSONContext;
 import com.euph28.tson.core.keyword.KeywordType;
-import com.euph28.tson.reporter.Report;
-import com.euph28.tson.reporter.ReportType;
+import com.euph28.tson.reporter.report.Report;
+import com.euph28.tson.reporter.report.ReportType;
 import com.euph28.tson.reporter.TSONReporter;
 import com.euph28.tson.restclientinterface.TSONRestClient;
